+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 <!-- and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). -->
 
+## Unreleased
+### Fixed
+- `submit/1` no longer relies on pressing Enter, so forms submit consistently when the last interacted element is a select.
+
 ## [0.14.0] 2026-05-05
 ### Added
 - Support phoenix_test [v0.11.1](https://hexdocs.pm/phoenix_test/changelog.html#0-11-1). Commit [f60d668]